requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.2
Django==3.2.23
gunicorn==21.2.0
idna==3.4
numpy==1.21.6
packaging==23.2
pandas==1.3.5
pickle5==0.0.11
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.7
